+++ backend/canisters/group/impl/src/updates/c2c_delete_group.rs
@@ -1,5 +1,6 @@
 use crate::{read_state, run_regular_jobs, RuntimeState};
 use canister_api_macros::update;
+use canister_client::make_c2c_call_raw;
 use canister_tracing_macros::trace;
 use group_canister::c2c_delete_group::{Response::*, *};
 use group_index_canister::c2c_delete_group;
@@ -22,10 +23,7 @@ async fn c2c_delete_group(_args: Args) -> Response {
         members: prepare_result.members,
     };
 
-    match group_index_canister_c2c_client::c2c_delete_group(group_index_canister_id, &c2c_delete_group_args).await {
-        Ok(_) => Success,
-        Err(error) => InternalError(format!("{error:?}")),
-    }
+    delete_group(group_index_canister_id, &c2c_delete_group_args).await
 }
 
 struct PrepareResult {
@@ -54,5 +52,18 @@ fn prepare(state: &RuntimeState) -> Result<PrepareResult, Response> {
         }
     } else {
         Err(NotAuthorized)
+    }
+}
+
+async fn delete_group(group_index_canister_id: CanisterId, args: &c2c_delete_group::Args) -> Response {
+    let method_name = "c2c_delete_group_msgpack";
+    let payload = msgpack::serialize_then_unwrap(args);
+    let c2c_cost = ic_cdk::api::cost_call(method_name.len() as u64, payload.len() as u64);
+    let buffer = 1_000_000_000; // 1B
+    let cycles = ic_cdk::api::canister_liquid_cycle_balance().saturating_sub(c2c_cost + buffer);
+
+    match make_c2c_call_raw(group_index_canister_id, method_name, &payload, cycles, None).await {
+        Ok(_) => Success,
+        Err(error) => InternalError(format!("{error:?}")),
     }
 }