+++ backend/canisters/group_index/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Increase timeout of bounded-wait calls to 5 minutes ([#7730](https://github.com/open-chat-labs/open-chat/pull/7730))
 - Include more details in failed c2c call errors ([#7749](https://github.com/open-chat-labs/open-chat/pull/7749))
+- Transfer cycles to GroupIndex when deleting a Group or Community ([#7768](https://github.com/open-chat-labs/open-chat/pull/7768))
 
 ### Fixed